server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
motor==3.3.1  # AsyncIO MongoDB driver
boto3==1.29.0
pydantic[email]==2.4.2
python-magic==0.4.27  # For file type detection